Got Teleporters Working
Teleporter Prefab takes transform of other GameObject as destination. When an object with 'Player' tag touches the teleporter they are moved to that postion.

Gets players movement and look scripts to disable them or the teleport. Without disabling the players position and rotation will not change. The names of these scripts are hardcoded into the portalScript so they will need to be changed

There are vertical and horizontal offset variables so the player can teleport in front of the object not in it. There is an arrow on the portal plate showing the forward direction of the portal