Fixed Room Position Bug and added Screen
fixed position bug but setting the rooms position in update() instead of start()

Added a screen in the main room that displays the view of a camera. This is done with a Render Texture. Set the camera's target texture to it and add it the screen to make it display