StaminaHUD and UpgradeData
Worked on StaminaHUD, to test I set keybinds to deplete and regenerate. The regenerate function is wrapped in a Coroutine so every press it takes about 2 seconds to heal but will modifty so its continuous or can just ignore that feature for MVP.

UpgradeData I referenced all the stats from testplayer ammo and stamina into upgradedata so I can call UnlockScripts inside of UpgradeData.