Upgrades Assigned to Room and Displayed
Rooms are now assigned upgrades from UpgradeSpace randomly

UpgradeSpace has a sample method that returns n random upgrades, without removing them from possibleUpgrades.

RoomManger gets a sample of 4  upgrades from UpgradeSpace and assigns each to a room. This is so that the upgrade will only get pulled from UpgradeSpace when a room is completed, and the upgrades assigned to unchosen rooms remain in possibleUpgrades

Text Displays for room now show the upgrade reward for completing the room