Fixed Disable Bug in Move and Look Scripts
Fixed bug where when Move and Look scripts were disabled and enabled they nolonger worked

The bug occured because the OnDisable() methods called disable() on each of the InputActions but in OnEnable() it did not reenable them. Even though they are reassigned in OnEnable() I guess the inputs are still disable for somereason.

I just commented out the disable() lines because i think they are redundant anyway