Actually Fixed the Bug For real this time!!!!
Fixed the telportation bug as well as another bug with enemies kills counting n times after clearing n rooms.

Bug 1: Player would not teleport before room deletion. This bug was actually the result of a bandiad fix for another bug i had before. When the room got delelted the player scripts would be disabled but the timer on that would reenable nolonger existed because it was deleted so the player coulnt move after returning to main. To fix this I made it so they reenabled the scripts OnDestroy, which fixed that issue but created this bug.

This bug happens because reenabling the scripts in OnDestroy defeats the purpose of disabling the scripts to begin with, which is stopping those scripts from overriding the teleport, which is what was happening. To fix this, I used the static timer I made for the telport lockout, so that the timer would increment after the portal is destroyed, and I made it so the script references are static so they can still be disabled.

Bug 2: After clearing n number of rooms, when killing an enemy, the numEnemies would decrement by n instead of by just 1. This bug happened because every RoomGeneration, the script was resubscribing to the portalEnter Event, so that ment that when a room was choosen, it was choosen n times, and therefore subscribed to the OnDeath event n times, which causes the number of enemies to decrement n times. To fix this, I moved the subscrition to the main room portals to the start method before generation, since this should only have to happen once and the portals in main shouldnt change.