Started writing unit tests for UpgradeSpace
wrote some test cases for looking up and pulling upgrades from UpgradeSpace so I can learn how. Right now I'm spending more of the time fixing my test cases because of silly errors, but hopefully ill get the hang of it and it shouldn't be that hard.

Uses a separate set of TestUpgrades to test